Define the minimum size required for the control structure - reduce memory footprint

+++ yabause/src/core/video/opengl/common/include/vidshared.h
@@ -194,8 +194,8 @@ typedef struct
 
    u16 char_bank[4];
    u16 pname_bank[4];
-   int alpha_per_line[270];
-   int draw_line;
+   u8 alpha_per_line[270];
+   u16 draw_line;
 
 } vdp2draw_struct;
